requirements.txt
Flask
Flask-CORS
google-generativeai
pdfplumber
python-docx
python-dotenv
